pyproject.toml
[project]
name = "stiflyt"
version = "0.1.0"
description = "Backend system for processing routes from turrutebasen and mapping matrikkelenhet"
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
    "psycopg2-binary>=2.9.9",
    "python-dotenv>=1.0.0",
    "fastapi>=0.104.0",
    "uvicorn[standard]>=0.24.0",
    "pydantic>=2.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "black>=23.0.0",
    "flake8>=6.0.0",
]

[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools]
packages = []

[tool.setuptools.package-data]
"*" = ["*.txt", "*.md"]

[tool.black]
line-length = 100
target-version = ['py39']

[tool.flake8]
max-line-length = 100
exclude = [".git", "__pycache__", "venv", ".venv"]

